requirements.txt
anyio==4.9.0
click==8.2.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Markdown==3.8
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pytest==8.3.5
python-frontmatter==1.1.0
PyYAML==6.0.2
ruff==0.11.11
sniffio==1.3.1
watchfiles==1.0.5
